Fix clippy warnings in katla_gfx
- Remove unused imports across Metal backend
- Remove unnecessary unsafe blocks (setLabel is safe in objc2)
- Fix doc comment formatting (empty line, list indentation)
- Fix field_reassign_with_default patterns via clippy --fix

+++ katla_gfx/src/metal/context.rs
@@ -106,7 +106,8 @@ pub(crate) fn default_pbr_vertex_descriptor() -> Retained<MTLVertexDescriptor> {
 /// - location 0: position Float2 @ offset 0
 /// - location 1: uv Float2 @ offset 8
 /// - location 2: color UByte4Norm @ offset 16
-/// (texture_index at offset 20 is not bound — passed via uniform instead)
+///
+/// texture_index at offset 20 is not bound — passed via uniform instead
 pub(crate) fn ui_vertex_descriptor() -> Retained<MTLVertexDescriptor> {
     let vertex_descriptor = MTLVertexDescriptor::new();
 
@@ -459,7 +460,7 @@ impl MetalContext {
 
         let color_attachments = descriptor.colorAttachments();
         for (i, &format) in color_formats.iter().enumerate() {
-            let attachment = unsafe { color_attachments.objectAtIndexedSubscript(i as usize) };
+            let attachment = unsafe { color_attachments.objectAtIndexedSubscript(i) };
             attachment.setPixelFormat(format);
 
             if alpha_blended {
@@ -603,7 +604,7 @@ impl MetalContext {
 
         let color_attachments = descriptor.colorAttachments();
         for (i, &format) in color_formats.iter().enumerate() {
-            let attachment = unsafe { color_attachments.objectAtIndexedSubscript(i as usize) };
+            let attachment = unsafe { color_attachments.objectAtIndexedSubscript(i) };
             attachment.setPixelFormat(format);
         }
